Move external resources tests to its own suite
