+++ test/acceptance/server_gettile.js
@@ -280,224 +280,6 @@ suite('server_gettile', function() {
         });
     });
 
-    test.skip("base and custom style tile referencing external resources do not affect each other",
-        function(done){
-      var style = "#test_table_3{marker-file: url('http://localhost:" + res_serv_port +
-          "/circle.svg'); marker-transform:'scale(0.2)'; }";
-      var style2 = "#test_table_3{marker-file: url('http://localhost:" + res_serv_port +
-          "/square.svg'); marker-transform:'scale(0.2)'; }";
-      var stylequery = querystring.stringify({style: style});
-      step(
-        function getCustomTile0() {
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?' + stylequery,
-            method: 'GET',
-            encoding: 'binary'
-          },{
-              status: 200,
-              headers: { 'Content-Type': 'image/png' }
-          }, function(res){
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg1.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  next(err);
-              });
-          });
-        },
-        // Set another style as default for table
-        function postStyle(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          //next(null, {statusCode: 200}); return;
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table_3/style',
-              method: 'POST',
-              headers: {'Content-Type': 'application/x-www-form-urlencoded' },
-              data: querystring.stringify({style: style2})
-          }, {}, function(res) { next(null, res); });
-        },
-        // Check style setting succeeded
-        function checkPost(err, res) {
-          if ( err ) {
-              throw err;
-          }
-          assert.equal(res.statusCode, 200, res.body);
-          //assert.equal(res.body, "ok");
-          return null;
-        },
-        // Now check we get the tile styled as we specified
-        function getBaseTile0(err/*, data*/) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png',
-            method: 'GET',
-            encoding: 'binary'
-          },{}, function(res){
-              assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg2.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  if (err) { next(err); return; }
-                  next(null);
-              });
-          });
-        },
-        // Now fetch the custom style tile again
-        function getCustomTile1(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?cache_buster=2&' + stylequery,
-            method: 'GET',
-            encoding: 'binary'
-          },{ }, function(res){
-              assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg1.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  next(err);
-              });
-          });
-        },
-        // Now fetch the base style tile again
-        function getBaseTile1(err/*, data*/) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?cache_buster=3',
-            method: 'GET',
-            encoding: 'binary'
-          },{ }, function(res){
-              assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg2.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  if (err) { next(err); return; }
-                  next(null);
-              });
-          });
-        },
-        function finish(err) {
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table_3/style',
-              method: 'DELETE' },{}, function() { done(err); });
-        }
-      );
-    });
-
-    // See http://github.com/CartoDB/Windshaft/issues/107
-    test.skip("external resources get localized on renderer creation",
-        function(done){
-      var style = "#test_table_3{marker-file: url('http://localhost:" + res_serv_port +
-          "/square.svg'); marker-transform:'scale(0.2)'; }";
-      var stylequery = querystring.stringify({style: style});
-      var numrequests;
-      step(
-        function getCustomTile0() {
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?cache_buster=2.1&' + stylequery,
-            method: 'GET',
-            encoding: 'binary'
-          },{
-              status: 200,
-              headers: { 'Content-Type': 'image/png' }
-          }, function(res){
-              numrequests = res_serv_status.numrequests;
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg2.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  next(err);
-              });
-          });
-        },
-        function dropLocalizedResources(err) {
-          if ( err ) {
-              throw err;
-          }
-          var cachedir = global.environment.millstone.cache_basedir;
-          rmdir_recursive_sync(cachedir);
-          // Reset server to ensure all renderer caches are flushed
-          server = new Windshaft.Server(ServerOptions);
-          server.setMaxListeners(0);
-          return null;
-        },
-        // Now fetch the custom style tile again
-        function getCustomTile1(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?cache_buster=2.2&' + stylequery,
-            method: 'GET',
-            encoding: 'binary'
-          },{}, function(res){
-              assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-              assert.equal(res_serv_status.numrequests, numrequests+1);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg2.png', IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  next(err);
-              });
-          });
-        },
-        // Now fetch the custom style tile again with an higher cache_buster,
-        // checking that the external resource is NOT downloaded again
-        function getCustomTile2(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-            url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?cache_buster=2.3&' + stylequery,
-            method: 'GET',
-            encoding: 'binary'
-          },{}, function(res){
-              assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-              // millstone should not make another request
-              assert.equal(res_serv_status.numrequests, numrequests+1);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_13_4011_3088_svg2.png',
-                  IMAGE_EQUALS_TOLERANCE_PER_MIL, function(err) {
-                  next(err);
-              });
-          });
-        },
-        function finish(err) {
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table_3/style',
-              method: 'DELETE' },{}, function() { done(err); });
-        }
-      );
-    });
-
-    test.skip("referencing unexistant external resources returns an error",
-        function(done){
-      var url = "http://localhost:" + res_serv_port + "/notfound.png";
-      var style = "#test_table_3{marker-file: url('" + url + "'); marker-transform:'scale(0.2)'; }";
-      var stylequery = querystring.stringify({style: style});
-      assert.response(server, {
-        url: '/database/windshaft_test/table/test_table_3/13/4011/3088.png?' + stylequery,
-        method: 'GET',
-        encoding: 'binary'
-      },{}, function(res){
-        assert.equal(res.statusCode, 400, res.body);
-        assert.equal(res.headers['content-type'], "application/json; charset=utf-8");
-        assert.deepEqual(JSON.parse(res.body), {
-            "error":"Unable to download '" + url + "' for 'style.mss' (server returned 404)"
-        });
-        done();
-      });
-    });
-
     // Test for https://github.com/Vizzuality/Windshaft/issues/65
     test.skip("catching non-Error exception doesn't kill the backend", function(done) {
